+++ api/views/camera_view.py
@@ -7,6 +7,8 @@ from api.models.camera_model import *
 from api.serializers import CameraSerializer
 from api.library.utils import splitHeader
 from bson import ObjectId
+from api.models.area_model import Area
+from api.models.role_model import Role
 
 
 class CameraView(APIView):
@@ -21,9 +23,13 @@ class CameraView(APIView):
         if bool(user) is False:
             return Response({'message': 'Authorization invalid.'}, status=status.HTTP_401_UNAUTHORIZED)
 
-        cameras = Camera.objects
-        camera_serializer = CameraSerializer(cameras, many=True)
-        return Response(camera_serializer.data, status=status.HTTP_200_OK)
+        user_role_login = Role.objects(id=user.role_id).first()
+        if user_role_login == 'admin':
+            cameras = Camera.objects
+            camera_serializer = CameraSerializer(cameras, many=True)
+            return Response(camera_serializer.data, status=status.HTTP_200_OK)
+
+        return Response({'message': 'Authorization invalid.'}, status=status.HTTP_401_UNAUTHORIZED)
 
     def post(self, request):
         if request.headers.get('Authorization') is None:
@@ -32,6 +38,10 @@ class CameraView(APIView):
             return Response({'message': 'Authorization invalid.'}, status=status.HTTP_401_UNAUTHORIZED)
         user = splitHeader(request.headers['Authorization'].split(' ')[1])
         if bool(user) is False:
+            return Response({'message': 'Authorization invalid.'}, status=status.HTTP_401_UNAUTHORIZED)
+
+        user_role_login = Role.objects(id=user.role_id).first()
+        if user_role_login.name == 'staff':
             return Response({'message': 'Authorization invalid.'}, status=status.HTTP_401_UNAUTHORIZED)
 
         camera_serializer = CameraSerializer(data=request.data)
@@ -79,6 +89,10 @@ class CameraDetailView(APIView):
         if bool(user) is False:
             return Response({'message': 'Authorization invalid.'}, status=status.HTTP_401_UNAUTHORIZED)
 
+        user_role_login = Role.objects(id=user.role_id).first()
+        if user_role_login.name == 'staff':
+            return Response({'message': 'Authorization invalid.'}, status=status.HTTP_401_UNAUTHORIZED)
+
         camera = Camera.objects(id=id).first()
         camera_serializer = CameraSerializer(camera, data=request.data, partial=True)
         if camera_serializer.is_valid():
@@ -98,14 +112,26 @@ class CameraDetailView(APIView):
         user = splitHeader(request.headers['Authorization'].split(' ')[1])
         if bool(user) is False:
             return Response({'message': 'Authorization invalid.'}, status=status.HTTP_401_UNAUTHORIZED)
-        camera = Camera.objects(id=id).first()
-        camera.delete()
-        return Response({'message': 'deleted successfully.'}, status=status.HTTP_200_OK)
+
+        try:
+            user_role_login = Role.objects(id=user.role_id).first()
+            if user_role_login.name == 'staff':
+                return Response({'message': 'Authorization invalid.'}, status=status.HTTP_401_UNAUTHORIZED)
+
+            # handle logic
+            camera = Camera.objects(id=id).first()
+            area = Area.objects(id=camera.area_id).first()
+            if user_role_login.name == 'manager' and str(user.id) != area.managed_manager:
+                return Response({'message': 'Authorization invalid.'}, status=status.HTTP_401_UNAUTHORIZED)
+            camera.delete()
+            return Response({'message': 'deleted successfully.'}, status=status.HTTP_200_OK)
+        except Exception as e:
+            print(e)
+            return Response({'message': 'object not found.'}, status=status.HTTP_401_UNAUTHORIZED)
 
 
 @api_view(['GET'])
 def getAllCamera(request: Request, area_id):
-    print(1)
     if request.headers.get('Authorization') is None:
         return Response({'message': 'Authorization invalid.'}, status=status.HTTP_401_UNAUTHORIZED)
     elif request.headers.get('Authorization').find('Bearer') == -1:
@@ -118,7 +144,7 @@ def getAllCamera(request: Request, area_id):
         return Response({'message': 'Authorization invalid.'}, status=status.HTTP_401_UNAUTHORIZED)
     try:
         offset = 0
-        limit_record = 10
+        limit_record = 0
         page_number = 1
 
         if bool(request.query_params.get('limit')) is True:
@@ -132,6 +158,18 @@ def getAllCamera(request: Request, area_id):
                 offset = (int(page_number) - 1) * int(limit_record)
             else:
                 return Response({'message': 'query page invalid'}, status=status.HTTP_400_BAD_REQUEST)
+
+        user_role_login = Role.objects(id=user.role_id).first()
+
+        if user_role_login.name == 'manager':
+            area = Area.objects(id=area_id).first()
+            print(area.managed_manager)
+            if area.managed_manager != str(user.id):
+                return Response({'message': 'Authorization invalid.'}, status=status.HTTP_401_UNAUTHORIZED)
+        elif user_role_login.name == 'staff':
+            area = Area.objects(id=area_id).first()
+            if area.managed_staff != str(user.id):
+                return Response({'message': 'Authorization invalid.'}, status=status.HTTP_401_UNAUTHORIZED)
 
         list_camera = Camera.objects(area_id=area_id).skip(offset).limit(int(limit_record))
         list_camera_serializer = CameraSerializer(list_camera, many=True)
